+++ common/src/main/java/org/geogebra/common/export/pstricks/GeoGebraToPdf.java
@@ -611,15 +611,15 @@ public abstract class GeoGebraToPdf extends GeoGebraExport {
 		double rwHeight = geo.getValue() * slopeTriangleSize;
 		double height = euclidianView.getYscale() * rwHeight;
 		double[] coords = new double[2];
-		if (Math.abs(height) > Float.MAX_VALUE) {
+		if (Math.abs(height) > Double.MAX_VALUE) {
 			return;
 		}
 		// get point on line g
 		((AlgoSlope) geo.getParentAlgorithm()).getInhomPointOnLine(coords);
 		// draw slope triangle
-		float x = (float) coords[0];
-		float y = (float) coords[1];
-		float xright = x + slopeTriangleSize;
+		double x = coords[0];
+		double y = coords[1];
+		double xright = x + slopeTriangleSize;
 		startBeamer(codeFilledObject);
 		codeFilledObject.append("\\draw");
 		String s = lineOptionCode(geo, true);
@@ -633,8 +633,8 @@ public abstract class GeoGebraToPdf extends GeoGebraExport {
 		writePoint(xright, y + rwHeight, codeFilledObject);
 		codeFilledObject.append(";\n");
 		// draw Label
-		float xLabelHor = (x + xright) / 2;
-		float yLabelHor = y - (float) ((euclidianView.getFont().getSize() + 2)
+		double xLabelHor = (x + xright) / 2;
+		double yLabelHor = y - ((euclidianView.getFont().getSize() + 2)
 				/ euclidianView.getYscale());
 		GColor geocolor = geo.getObjectColor();
 		codePoint.append("\\draw[color=");
@@ -986,7 +986,7 @@ public abstract class GeoGebraToPdf extends GeoGebraExport {
 	@Override
 	protected void drawPolygon(GeoPolygon geo) {
 		// command: \pspolygon[par](x0,y0)....(xn,yn)
-		float alpha = geo.getAlphaValue();
+		double alpha = geo.getAlphaValue();
 		if (alpha == 0.0f && geo.getFillType() == FillType.IMAGE)
 			return;
 		startBeamer(code);
